requirements.txt
Django = 2.2.6
djangorestframework
markdown
django-filter
